+++ PsychSourceGL/Source/Common/Screen/SCREENOpenMovie.c
@@ -414,8 +414,8 @@ PsychError SCREENCreateMovie(void)
 	
 	// Get the optional size:
 	// Default Width and Height of movie frames is derived from size of window:
-	width = PsychGetWidthFromRect(windowRecord->clientrect);
-	height = PsychGetHeightFromRect(windowRecord->clientrect);
+	width  = (int) PsychGetWidthFromRect(windowRecord->clientrect);
+	height = (int) PsychGetHeightFromRect(windowRecord->clientrect);
 	PsychCopyInIntegerArg(3, kPsychArgOptional, &width);
 	PsychCopyInIntegerArg(4, kPsychArgOptional, &height);
